created serverUtil file, added validateCartItems there
